Do not return Cohort / IR / PLE / PLP results for sources which a user doesn't have access to (#857)
* Do not return Cohort / IR / PLE / PLP results for sources which a user doesn't have access to

+++ src/main/java/org/ohdsi/webapi/executionengine/service/ScriptExecutionService.java
@@ -18,5 +18,5 @@ public interface ScriptExecutionService {
 
     String getExecutionStatus(Long executionId);
 
-    List<AnalysisResultFile> getExecutionResultFiles(Long executionId);
+    List<AnalysisResultFile> getExecutionResultFiles(Integer executionId);
 }
